rename wait-for-data.js to makeRequests.js relates > #52
